Drag and Drop Refactor - Part 1
First step in refactoring the messy and mostly unusable drag and drop
system. This introduces the use of target lists, which is a much more
elegant way of handling targets. It also properly uses the target
system now that I know how they actually work.

There are some regressions with this implementation:
 - External drag and drop support is not yet implemented, as we
   need to indicate that we prefer the BlDocument target where
   possible (otherwise untitled files will not work).
 - Secondly, when dragging and dropping an untitled file, we want to
   disallow the URI-list target, as it has not been saved. We should
   allow plain text drag-and-drop however.
 - Finally, the font-selection system is completely broken, having
   been modified for a slight BlDocument refactor. This will probably
   need to be redone from scratch, along with a properties dialogue.

Signed-off-by: Matthew Jakeman <mjakeman26@outlook.co.nz>