Preferences Second Iteration (Added libhandy + GSettings Integration)
Redesigned the Preferences window using HdyPreferences. The window
scales adaptively across screen sizes (which opens the possibility
for a simplified mobile version, or touch screen support). Libhandy
was always intended to be a dependency so I feel alright adding it
as a hard requirement.

Additionally, GSettings integration has been added to the preferences
window, with the data being synced to/from the backend. The current
implementation is very verbose, and there are probably better ways to
go about doing this, but it suffices for now.

The next step will be to update BlMarkdownView to follow these
preferences. While it might be an idea to allow for overriding
default settings in the future, an application wide approach will be
better to start with.

Signed-off-by: Matthew Jakeman <mjakeman26@outlook.co.nz>